+++ cucumber-junit-platform-engine/src/main/java/io/cucumber/junit/platform/engine/CucumberEngineOptions.java
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static io.cucumber.core.resource.ClasspathSupport.CLASSPATH_SCHEME_PREFIX;
 import static io.cucumber.junit.platform.engine.Constants.ANSI_COLORS_DISABLED_PROPERTY_NAME;
@@ -71,41 +70,39 @@ class CucumberEngineOptions implements
     }
 
     private Optional<PluginOption> getPublishPlugin() {
-        Optional<PluginOption> fromToken = getPublishTokenPlugin();
-        Optional<PluginOption> fromEnabled = getPublishEnabledPlugin();
-
-        Optional<PluginOption> plugin = Stream.of(fromToken, fromEnabled)
-                .flatMap(pluginOption -> pluginOption.map(Stream::of).orElseGet(Stream::empty))
-                .findFirst();
-
-        // With higher java version use ifPresentOrElse in plugins()
-        if (plugin.isPresent()) {
-            return plugin;
+        if (isPublishPluginEnabled()) {
+            return createPublishPlugin();
         }
-        return getPublishQuietPlugin();
+        return createCucumberReportsAdvertisingPlugin();
     }
 
-    private Optional<PluginOption> getPublishQuietPlugin() {
+    private Optional<PluginOption> createCucumberReportsAdvertisingPlugin() {
         Optional<PluginOption> noPublishOption = Optional.of(PluginOption.forClass(NoPublishFormatter.class));
         Optional<PluginOption> quiteOption = Optional.empty();
         return configurationParameters
                 .getBoolean(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME)
                 .map(quite -> quite ? quiteOption : noPublishOption)
-                .orElse(noPublishOption);
+                // Disable the banner advertising the hosted cucumber reports
+                // by default until the uncertainty around the projects future
+                // is resolved. It would not be proper to advertise a service
+                // that may be discontinued to new users.
+                // For context see: https://mattwynne.net/new-beginning
+                .orElse(quiteOption);
+
     }
 
-    private Optional<PluginOption> getPublishTokenPlugin() {
-        return configurationParameters
+    private Optional<PluginOption> createPublishPlugin() {
+        PluginOption publishPlugin = configurationParameters
                 .get(PLUGIN_PUBLISH_TOKEN_PROPERTY_NAME)
-                .map(token -> PluginOption.forClass(PublishFormatter.class, token));
+                .map(token -> PluginOption.forClass(PublishFormatter.class, token))
+                .orElse(PluginOption.forClass(PublishFormatter.class));
+        return Optional.of(publishPlugin);
     }
 
-    private Optional<PluginOption> getPublishEnabledPlugin() {
-        Optional<PluginOption> enabledOption = Optional.of(PluginOption.forClass(PublishFormatter.class));
-        Optional<PluginOption> disabledOption = Optional.empty();
-        return configurationParameters
-                .getBoolean(PLUGIN_PUBLISH_ENABLED_PROPERTY_NAME)
-                .flatMap(enabled -> enabled ? enabledOption : disabledOption);
+    private boolean isPublishPluginEnabled() {
+        return configurationParameters.getBoolean(PLUGIN_PUBLISH_ENABLED_PROPERTY_NAME)
+                // Implicitly enabled by the token if not explicitly disabled
+                .orElse(configurationParameters.get(PLUGIN_PUBLISH_TOKEN_PROPERTY_NAME).isPresent());
     }
 
     @Override